backend/requirements.txt
fastapi
uvicorn
google-genai
google-adk
google-cloud-logging
python-multipart
playwright
Pillow
